+++ Extension/src/background/api/filters/userrules.ts
@@ -17,7 +17,8 @@
  * You should have received a copy of the GNU General Public License
  * along with AdGuard Browser Extension. If not, see <http://www.gnu.org/licenses/>.
  */
-import { FilterList, RuleSyntaxUtils } from '@adguard/tsurlfilter';
+import { FilterList } from '@adguard/tswebextension';
+import { RuleSyntaxUtils } from '@adguard/tsurlfilter';
 
 import { logger } from '../../../common/logger';
 import {